work in progress on working with nested tibble
This rewrites collapse_cycles to nest the results by file_id and calculate
summary statistics as new columns. This should be faster, and still allow us to
hack back into the raw results calculated before.

Present problem is that when I merge it with the metadata from
iso_get_file_info the measurement_info list column cannot be unnested.

Todo: use the same approach for the ETF-calculation. See also @sebkopf's
comment in #9.